Fixed 32bit build errors

+++ zmq.rs
@@ -2,7 +2,7 @@
 Module: zmq
 */
 
-import libc::{c_int, c_long, c_void, size_t, c_char};
+import libc::{c_int, c_long, c_longlong, c_void, size_t, c_char};
 import result::{result, ok, err, chain};
 
 export context;
@@ -85,7 +85,7 @@ native mod zmq {
     fn zmq_send(socket: socket, msg: msg, flags: c_int) -> c_int;
     fn zmq_recv(socket: socket, msg: msg, flags: c_int) -> c_int;
 
-    fn zmq_poll(items: *pollitem, nitems: c_int, timeout: c_long) -> c_int;
+    fn zmq_poll(items: *pollitem, nitems: c_int, timeout: c_longlong) -> c_int;
 }
 
 #[doc = "Socket types"]
@@ -286,7 +286,7 @@ impl socket for socket {
         }
     }
 
-    fn get_fd() -> result<int, error> {
+    fn get_fd() -> result<i64, error> {
         getsockopt_i64(self, constants::ZMQ_FD)
     }
